router uses postmessage
