--- v0.1/deno/router_types.ts
@@ -1,17 +1,0 @@
-// brian taylor vann
-
-type HistoryModifier =
-  | "router_broadcast"
-  | "router_broadcast_unknown"
-  | "router_broadcast_hash_change";
-
-type BroadcastMessage<D = unknown> = {
-  type: HistoryModifier;
-  location: string;
-  title: string;
-  data?: D;
-};
-
-type Callback = <D>(data: BroadcastMessage<D>) => void;
-
-export type { BroadcastMessage, Callback, HistoryModifier };
