--- v0.1/deno/utils.ts
@@ -1,25 +1,0 @@
-import type { BroadcastMessage, HistoryModifier } from "./router_types.ts";
-
-const BROADCAST = "router_broadcast";
-const HASH_CHANGE = "router_broadcast_hash_change";
-const PAGE_SHOW = "router_broadcast_unknown";
-
-function getLocation(): string {
-  return window.location.href.substring(window.origin.length);
-}
-
-function replaceHistoryEntry<D>(type: HistoryModifier) {
-  const location = getLocation();
-  const { title } = document;
-
-  const state: BroadcastMessage<D> = {
-    data: history.state?.data,
-    type,
-    location,
-    title,
-  };
-
-  history.replaceState(state, title, location);
-}
-
-export { BROADCAST, getLocation, HASH_CHANGE, PAGE_SHOW, replaceHistoryEntry };
