+++ v0.1/es/router.js
@@ -1,6 +1,9 @@
+const POPSTATE = "popstate";
+const PAGESHOW = "pageshow";
 const BROADCAST = "router_broadcast";
-const HASH_CHANGE = "router_broadcast_hash_change";
-const PAGE_SHOW = "router_broadcast_unknown";
+const HASH_CHANGE = "router_hash_change";
+const UNKNOWN = "router_unknown";
+const bc = BroadcastChannel ? new BroadcastChannel("router") : window;
 function getLocation() {
   return window.location.href.substring(window.origin.length);
 }
@@ -15,28 +18,24 @@ function replaceHistoryEntry(type) {
   };
   history.replaceState(state, title, location);
 }
-const POPSTATE = "popstate";
-const PAGESHOW = "pageshow";
-let callback = () => {
-};
-function subscribe(cb) {
-  callback = cb;
+function push(state) {
+  const { title, location } = state;
+  history.pushState(state, title, location);
+  bc.postMessage(history.state);
 }
 function onPopState(e) {
-  console.log("on pop state");
-  if (e.state === null) {
-    replaceHistoryEntry(HASH_CHANGE);
-  }
-  callback(history.state);
+  if (e.state === null) replaceHistoryEntry(HASH_CHANGE);
+  bc.postMessage(history.state);
 }
 function onPageShow() {
-  console.log("on page show");
-  if (history.state === null) {
-    replaceHistoryEntry(PAGE_SHOW);
-  }
-  callback(history.state);
+  if (history.state === null) replaceHistoryEntry(UNKNOWN);
+  bc.postMessage(history.state);
 }
-window.addEventListener(PAGESHOW, onPageShow);
 window.addEventListener(POPSTATE, onPopState);
-export { subscribe as subscribe };
-export { BROADCAST as BROADCAST, HASH_CHANGE as HASH_CHANGE };
+window.addEventListener(PAGESHOW, onPageShow);
+export {
+  BROADCAST as BROADCAST,
+  HASH_CHANGE as HASH_CHANGE,
+  push as push,
+  UNKNOWN as UNKNOWN,
+};
